adding gizmo mode menu to unity state and taking properties from there, already working.
