+++ Assets/Scripts/Interaction/GeneralInputInteraction.cs
@@ -131,7 +131,7 @@ namespace Assets.Scripts.Interaction
             Ray mouseRay;
             bool isMouseOverGameWindow;
             bool isMouseOverContextMenu;
-            //bool isMouseOverGizmoModeMenu;
+            bool isMouseOverGizmoModeMenu;
             {
                 var mousePosViewport = inputHelper.GetMousePositionInScenePanel();
                 // Get the ray from the Camera, that corresponds to the mouse position in the panel
@@ -142,7 +142,7 @@ namespace Assets.Scripts.Interaction
                                         mousePosViewport.y >= 0 &&
                                         mousePosViewport.y < 1;
                 isMouseOverContextMenu = contextMenuSystem.IsMouseOverMenu();
-                //isMouseOverGizmoModeMenu = mouseHoverSceneButton.IsMouseOverGizmoModeMenu;
+                isMouseOverGizmoModeMenu = unityState.GizmoModeMenu.GetComponent<GizmoModeInteraction>().IsMouseOverGizmoModeMenu;
             }
 
 
@@ -150,7 +150,7 @@ namespace Assets.Scripts.Interaction
 
 
             // Figure out, if the mouse is over the 3D viewport (not hovering over any UI)
-            var isMouseIn3DView = /*!EventSystem.current.IsPointerOverGameObject() &&*/ isMouseOverGameWindow && !isMouseOverContextMenu; //&& !isMouseOverGizmoModeMenu;
+            var isMouseIn3DView = /*!EventSystem.current.IsPointerOverGameObject() &&*/ isMouseOverGameWindow && !isMouseOverContextMenu && !isMouseOverGizmoModeMenu;
 
             // The position the mouse currently points to in the 3D viewport
             Vector3? mousePositionIn3DView = null;
