--- Assets/Scripts/Interaction/MouseHoverSceneButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 19ebb3b7f453e984d9ed994bf2a17c13
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
